+++ R/fit_survival_step.R
@@ -23,7 +23,6 @@
 #' # Testing dataset with just two treatment arms.
 #' library(random.cdisc.data)
 #' library(dplyr)
-#' library(forcats)
 #'
 #' adtte <- radtte(cached = TRUE)
 #'
@@ -34,7 +33,7 @@
 #'   ) %>%
 #'   mutate(
 #'     # Reorder levels of ARM to display reference arm before treatment arm.
-#'     ARM = droplevels(fct_relevel(ARM, "B: Placebo")),
+#'     ARM = droplevels(forcats::fct_relevel(ARM, "B: Placebo")),
 #'     is_event = CNSR == 0
 #'   ) %>%
 #'   var_relabel(