+++ R/g_step.R
@@ -64,12 +64,15 @@ tidy.step <- function(x, ...) {  #nolint #nousage
 #' @examples
 #' library(survival)
 #' lung$sex <- factor(lung$sex)
+#'
+#' # Survival example.
 #' vars <- list(
 #'   time = "time",
 #'   event = "status",
 #'   arm = "sex",
 #'   biomarker = "age"
 #' )
+#'
 #' step_matrix <- fit_survival_step(
 #'   variables = vars,
 #'   data = lung,
@@ -98,6 +101,24 @@ tidy.step <- function(x, ...) {  #nolint #nousage
 #' step_data$extra <- exp(step_data$`Percentile Center`)
 #' g_step(step_data) +
 #'   geom_line(aes(y = extra), linetype = 2, color = "green")
+#'
+#' # Response example.
+#' vars <- list(
+#'   response = "status",
+#'   arm = "sex",
+#'   biomarker = "age"
+#' )
+#'
+#' step_matrix <- fit_rsp_step(
+#'   variables = vars,
+#'   data = lung,
+#'   control = c(
+#'     control_logistic(response_definition = "I(response == 2)"),
+#'     control_step()
+#'   )
+#' )
+#' step_data <- broom::tidy(step_matrix)
+#' g_step(step_data)
 #'
 g_step <- function(df,
                    use_percentile = "Percentile Center" %in% names(df),