+++ tests/testthat/test-fit_rsp_step.R
@@ -1,0 +1,91 @@
+library(random.cdisc.data)
+
+get_data <- function() {
+  adrs <- radrs(cached = TRUE)
+  adrs %>%
+    dplyr::filter(
+      PARAMCD == "BESRSPI",
+      ARM %in% c("B: Placebo", "A: Drug X")
+    ) %>%
+    dplyr::mutate(
+      # Reorder levels of ARM to have Placebo as reference arm for Odds Ratio calculations.
+      ARM = droplevels(forcats::fct_relevel(ARM, "B: Placebo")),
+      RSP = dplyr::case_when(AVALC %in% c("PR", "CR") ~ 1, TRUE ~ 0),
+      SEX = factor(SEX)
+    )
+}
+
+test_that("fit_rsp_step works as expected with default options", {
+  data <- get_data()
+  variables <- list(
+    arm = "ARM",
+    biomarker = "BMRKR1",
+    covariates = "AGE",
+    response = "RSP"
+  )
+  result <- expect_silent(fit_rsp_step(
+    variables = variables,
+    data = data
+  ))
+  expect_is(result, c("matrix", "step"))
+  expect_identical(ncol(result), 11L)
+  expect_identical(
+    colnames(result),
+    c("Percentile Center", "Percentile Lower", "Percentile Upper",
+      "Interval Center", "Interval Lower", "Interval Upper", "n",
+      "logor", "se", "ci_lower", "ci_upper")
+  )
+})
+
+test_that("fit_rsp_step works as expected with global model fit", {
+  data <- get_data()
+  variables <- list(
+    arm = "ARM",
+    biomarker = "BMRKR1",
+    covariates = "AGE",
+    response = "RSP"
+  )
+  result <- expect_silent(fit_rsp_step(
+    variables = variables,
+    data = data,
+    control = c(
+      control_logistic(conf_level = 0.8),
+      control_step(
+        biomarker = data$BMRKR1,
+        use_percentile = FALSE,
+        num_points = 3L
+      )
+    )
+  ))
+  expect_is(result, c("matrix", "step"))
+  expect_identical(ncol(result), 8L)
+  expect_identical(
+    colnames(result),
+    c("Interval Center", "Interval Lower", "Interval Upper", "n",
+      "logor", "se", "ci_lower", "ci_upper")
+  )
+})
+
+test_that("fit_rsp_step works as expected with strata", {
+  data <- get_data()
+  variables <- list(
+    arm = "ARM",
+    biomarker = "BMRKR1",
+    covariates = "AGE",
+    response = "RSP",
+    strata = c("STRATA1", "STRATA2")
+  )
+  result <- expect_silent(fit_rsp_step(
+    variables = variables,
+    data = data,
+    control = c(control_logistic(), control_step(bandwidth = 0.6))
+  ))
+  expect_is(result, c("matrix", "step"))
+  expect_identical(ncol(result), 11L)
+  expect_identical(
+    colnames(result),
+    c("Percentile Center", "Percentile Lower", "Percentile Upper",
+      "Interval Center", "Interval Lower", "Interval Upper", "n",
+      "logor", "se", "ci_lower", "ci_upper")
+  )
+})