+++ tests/testthat/test-h_step.R
@@ -1,3 +1,5 @@
+library(random.cdisc.data)
+
 get_simple <- function() {
   data.frame(
     time = c(5, 5, 10, 10, 5, 5, 5, 10, 10, 5),
@@ -7,8 +9,18 @@ get_simple <- function() {
     stage = factor(
       c("1", "2", "1", "1", "1", "1", "2", "1", "2", "1"),
       levels = c("1", "2")
-    )
+    ),
+    rsp = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE)
   )
+}
+
+get_adrs <- function() {
+  radrs(cached = TRUE) %>%
+    dplyr::filter(ARMCD %in% c("ARM A", "ARM B")) %>%
+    dplyr::mutate(
+      RSP = dplyr::case_when(AVALC %in% c("PR", "CR") ~ 1, TRUE ~ 0),
+      ARMBIN = droplevels(ARMCD)
+    )
 }
 
 # h_step_window ----
@@ -147,13 +159,7 @@ test_that("h_step_trt_effect works for logistic regression models without intera
 })
 
 test_that("h_step_trt_effect works for logistic regression models with interaction", {
-  library(random.cdisc.data)
-  dta <- radrs(cached = TRUE) %>%
-    dplyr::filter(ARMCD %in% c("ARM A", "ARM B")) %>%
-    dplyr::mutate(
-      RSP = dplyr::case_when(AVALC %in% c("PR", "CR") ~ 1, TRUE ~ 0),
-      ARMBIN = droplevels(ARMCD)
-    )
+  dta <- get_adrs()
   mod <- glm(formula = RSP ~ ARMBIN * AGE + RACE, data = dta, family = binomial())
   vars <- list(
     arm = "ARMBIN",
@@ -179,6 +185,42 @@ test_that("h_step_trt_effect works for logistic regression models with interacti
   expect_equivalent(result["est"], est_manual)
   # Regression test for se.
   expect_equivalent(result["se"], 0.2153840, tol = 1e-5)
+})
+
+test_that("h_step_trt_effect works for conditional logistic regression without interaction", {
+  dta <- get_adrs()
+  mod <- clogit(formula = RSP ~ ARMBIN + strata(STRATA1), data = dta)
+  vars <- list(
+    arm = "ARMBIN",
+    biomarker = "AGE"
+  )
+  result <- h_step_trt_effect(
+    data = dta,
+    model = mod,
+    variables = vars,
+    x = 0
+  )
+  coef_sum <- summary(mod)$coefficients
+  expect_equivalent(result, coef_sum["ARMBINARM B", c("coef", "se(coef)")])
+})
+
+test_that("h_step_trt_effect works for conditional logistic regression with interaction", {
+  dta <- get_adrs()
+  mod <- clogit(formula = RSP ~ ARMBIN * AGE + strata(STRATA1), data = dta)
+  vars <- list(
+    arm = "ARMBIN",
+    biomarker = "AGE"
+  )
+  result <- h_step_trt_effect(
+    data = dta,
+    model = mod,
+    variables = vars,
+    x = 50
+  )
+  est_manual <- sum(coef(mod)[c("ARMBINARM B", "AGE", "ARMBINARM B:AGE")] * (c(1, 50, 50) - c(0, 50, 0)))
+  expect_equivalent(result["est"], est_manual)
+  # Regression test for se.
+  expect_equivalent(result["se"], 0.2100507, tol = 1e-5)
 })
 
 # h_step_survival_formula ----
@@ -245,4 +287,121 @@ test_that("h_step_survival_est works as expected", {
   expect_identical(dim(result), c(3L, 6L))
   expect_identical(colnames(result), c("n", "events", "loghr", "se", "ci_lower", "ci_upper"))
   expect_equal(result[, "loghr"], c(1.075486, 1.082507, 1.089528), tol = 1e-6)
+})
+
+# h_step_rsp_formula ----
+
+test_that("h_step_rsp_formula works correctly without covariates", {
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "RSP"),
+      control = c(control_step(degree = 3), control_logistic())
+    ),
+    RSP ~ TRT * poly(BM, degree = 3, raw = TRUE)
+  )
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "RSP")
+    ),
+    RSP ~ TRT
+  )
+})
+
+test_that("h_step_rsp_formula works correctly with covariates", {
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "RSP", covariates = c("A", "B")),
+      control = c(control_logistic(), control_step(degree = 2))
+    ),
+    RSP ~ TRT * poly(BM, degree = 2, raw = TRUE) + A + B
+  )
+})
+
+test_that("h_step_rsp_formula works correctly with different response definition", {
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "AVAL"),
+      control = c(
+        control_logistic(response_definition = "I(response == 'CR')"),
+        control_step(degree = 1)
+      )
+    ),
+    I(AVAL == "CR") ~ TRT * poly(BM, degree = 1, raw = TRUE)
+  )
+})
+
+test_that("h_step_rsp_formula works correctly with strata", {
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "RSP", strata = c("A", "B")),
+      control = c(control_logistic(), control_step(degree = 2))
+    ),
+    RSP ~ TRT * poly(BM, degree = 2, raw = TRUE) + strata(I(interaction(A, B)))
+  )
+  expect_equal(
+    h_step_rsp_formula(
+      list(arm = "TRT", biomarker = "BM", response = "RSP", strata = "A"),
+      control = c(control_logistic(), control_step(degree = 2))
+    ),
+    RSP ~ TRT * poly(BM, degree = 2, raw = TRUE) + strata(A)
+  )
+})
+
+# h_step_rsp_est ----
+
+test_that("h_step_rsp_est works as expected without strata", {
+  dta_simple <- get_simple()
+  vars <- list(
+    arm = "armcd",
+    biomarker = "age",
+    response = "rsp",
+    covariates = "time"
+  )
+  age_vals <- c(20, 30, 40)
+  form <- h_step_rsp_formula(
+    variables = vars,
+    control = c(control_logistic(), control_step(degree = 1))
+  )
+  subset <- dta_simple$age > 15
+  result <- expect_silent(h_step_rsp_est(
+    formula = form,
+    data = dta_simple,
+    variables = vars,
+    subset = subset,
+    x = age_vals
+  ))
+  expect_is(result, "matrix")
+  expect_identical(dim(result), c(3L, 5L))
+  expect_identical(colnames(result), c("n", "logor", "se", "ci_lower", "ci_upper"))
+  expect_equal(result[, "logor"], c(2.008012, 1.025773, 0.043535), tol = 1e-6)
+  expect_equal(result[, "n"], rep(sum(subset), 3L))
+})
+
+test_that("h_step_rsp_est works as expected with strata", {
+  dta_simple <- get_simple()
+  vars <- list(
+    arm = "armcd",
+    biomarker = "age",
+    response = "rsp",
+    covariates = "time",
+    strata = "stage"
+  )
+  age_vals <- c(20, 30, 40)
+  form <- h_step_rsp_formula(
+    variables = vars,
+    control = c(control_logistic(), control_step(degree = 1))
+  )
+  subset <- dta_simple$age > 15
+  result <- expect_silent(h_step_rsp_est(
+    formula = form,
+    data = dta_simple,
+    variables = vars,
+    subset = subset,
+    x = age_vals
+  ))
+  expect_is(result, "matrix")
+  expect_identical(dim(result), c(3L, 5L))
+  expect_identical(colnames(result), c("n", "logor", "se", "ci_lower", "ci_upper"))
+  expect_equal(result[, "logor"], c(1.134057, 0.511102, -0.111853), tol = 1e-6)
+  expect_equal(result[, "n"], rep(sum(subset), 3L))
 })